fixed data types

+++ src/main/java/Main.java
@@ -2,7 +2,7 @@ import ru.netology.stats.StatsService;
 
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
         StatsService service = new StatsService();
