+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -8,9 +8,9 @@ public class StatsServiceTest {
     @Test
     public void checkSalesSum() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 180;
-        int actual = service.salesSum(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 180;
+        long actual = service.salesSum(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -18,9 +18,9 @@ public class StatsServiceTest {
     @Test
     public void checkSalesAvg() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 15;
-        int actual = service.salesAvg(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 15;
+        long actual = service.salesAvg(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -28,7 +28,7 @@ public class StatsServiceTest {
     @Test
     public void checkMaxLastSaleMonthNumber() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 8;
         int actual = service.maxSalesMonthIndexLast(sales);
 
@@ -38,7 +38,7 @@ public class StatsServiceTest {
     @Test
     public void checkMinLastSaleMonthNumber() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 9;
         int actual = service.minSalesMonthIndexLast(sales);
 
@@ -48,7 +48,7 @@ public class StatsServiceTest {
     @Test
     public void checkBelowAvgSalesMonths() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
         int actual = service.belowAvgSalesMonths(sales);
 
@@ -58,7 +58,7 @@ public class StatsServiceTest {
     @Test
     public void checkAboveAvgSalesMonths() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
         int actual = service.aboveAvgSalesMonths(sales);
